+++ src/SGDE.Domain/Entities/DetailEmbargo.cs
@@ -6,7 +6,7 @@ namespace SGDE.Domain.Entities
     {
         public DateTime DatePay { get; set; } 
         public string Observations { get; set; }
-        public Decimal Amount { get; set; }
+        public double Amount { get; set; }
 
         public int EmbargoId { get; set; }
         public virtual Embargo Embargo { get; set; }
