+++ src/SGDE.Domain/Entities/Training.cs
@@ -9,7 +9,7 @@
     public class Training : BaseEntity
     {
         public string Name { get; set; }
-        public Decimal Hours { get; set; }
+        public double Hours { get; set; }
         public string Center { get; set; }
         public string Address { get; set; }
         public byte[] File { get; set; }
