+++ src/SGDE.Domain/ViewModels/CostWorkerViewModel.cs
@@ -8,11 +8,11 @@
 
     public class CostWorkerViewModel : BaseEntityViewModel
     {
-        public Decimal priceHourOrdinary { get; set; }
-        public Decimal priceHourExtra { get; set; }
-        public Decimal priceHourFestive { get; set; }
-        public Decimal priceHourNocturnal { get; set; }
-        public Decimal priceDaily { get; set; }
+        public double priceHourOrdinary { get; set; }
+        public double priceHourExtra { get; set; }
+        public double priceHourFestive { get; set; }
+        public double priceHourNocturnal { get; set; }
+        public double priceDaily { get; set; }
         public string startDate { get; set; }
         public string endDate { get; set; }
         public string observations { get; set; }
